refactor(api): move projects to modules/ as controller, model, service
