+++ apps/api/src/issues/store.ts
@@ -33,7 +33,7 @@ import {
 } from 'drizzle-orm';
 import type { IssueQuery } from '#modules/agents/core/issue-query';
 import { iso, num, HttpError } from '../shared/lib';
-import type { ProjectRow } from '../projects/store';
+import type { ProjectRow } from '#modules/projects/service';
 import { getCustomFieldById, type CustomFieldType } from '#modules/custom-fields/service';
 import {
   recordActivity,
