+++ apps/api/src/modules/agents/core/runtime/tools/route-tools.ts
@@ -1,6 +1,6 @@
 import { createTool } from '@mastra/core/tools';
 import type { z } from 'zod';
-import type { ProjectRow } from '../../../../../projects/store';
+import type { ProjectRow } from '#modules/projects/service';
 import { routeTools, type McpInputSchema, type McpRouteTool } from '#mcp/generate';
 import { dispatchTool } from '#mcp/dispatch';
 import { getMcpApp } from '#mcp/app-ref';
