+++ apps/worker/src/store.ts
@@ -148,12 +148,13 @@ export async function cleanupOldDeliveries(): Promise<number> {
 // threshold lives in project_setting under key 'auto_archive' as
 // { completedDays, canceledDays }; a positive day count enables archiving for that
 // state group, null/absent disables it (kept in sync with getAutoArchiveSettings in
-// apps/api/src/projects/store.ts). Inactivity is measured by issue.updated_at:
-// moving to a terminal column bumps it, and any later edit resets the clock, so an
-// issue is archived only after the full period with no activity. The ->> is guarded
-// by a numeric-string regex so a malformed or missing value is treated as disabled,
-// never cast. Returns the ids archived, which the caller uses to drop the agent
-// conversation threads of those issues. Idempotent (archived_at IS NULL filter).
+// apps/api/src/modules/projects/service.ts). Inactivity is measured by
+// issue.updated_at: moving to a terminal column bumps it, and any later edit resets
+// the clock, so an issue is archived only after the full period with no activity.
+// The ->> is guarded by a numeric-string regex so a malformed or missing value is
+// treated as disabled, never cast. Returns the ids archived, which the caller uses
+// to drop the agent conversation threads of those issues. Idempotent
+// (archived_at IS NULL filter).
 export async function archiveStaleIssues(): Promise<number[]> {
   const rows = await db.execute(sql`
     UPDATE issue i
